Module Zarr_eio.Storage

module IO : Zarr.Types.IO with type 'a t = 'a

module MemoryStore : Zarr.Memory.S with type 'a io := 'a

An Eio-aware in-memory storage backend for Zarr v3 hierarchy.

module ZipStore : Zarr.Zip.S with type 'a io := 'a

An Eio-aware Zip file storage backend for a Zarr v3 hierarchy.

module FilesystemStore : sig ... end

An Eio-aware local filesystem storage backend for a Zarr v3 hierarchy.