Module Storage.ZipStore

An Eio-aware Zip file storage backend for a Zarr v3 hierarchy.

exception Path_already_exists of string

include Zarr.Storage.S with type 'a io := 'a

type t

The storage type.

module Group : sig ... end

module Array : sig ... end

val hierarchy : t -> Zarr.Node.Array.t list * Zarr.Node.Group.t list

hierarchy t returns p where p is a pair of lists representing all nodes in store t. The first element of the pair is a list of all array nodes, and the second element is a list of all group nodes. This operation returns a pair of empty lists if store t is empty.

raises Parse_error
if any node has invalid node_type metadata.

val clear : t -> unit

clear t clears the store t by deleting all nodes. If the store is already empty, this is a no-op.

val open_store : ?level:Zarr.Codecs.deflate_level -> string -> t

open_store ?level p returns a store instance representing a zip archive of a Zarr v3 hierarchy stored at path p. level is the DEFLATE algorithm compression setting used when writing new entries into the archive.

val create : ?level:Zarr.Codecs.deflate_level -> string -> t

create ?level p creates a zip archive at path p and then returns a store instance representing the zip archive. level is the DEFLATE algorithm compression setting used when writing new entries into the archive.

raises Path_already_exists
if a file already exists at path p.